+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 
+- In some case request handle error will result in uncompleted or invalid http response. [`#417`](https://github.com/project-vrcz/content-publisher/pull/417)
 - Task will always fail due to retry non-idempotent requests. [`#411`](https://github.com/project-vrcz/content-publisher/pull/411)
   - (caused by timeout or connection abort during received response)
 - In some case app crash will corrupt settings file.